requirements.txt
numpy==2.2.5
pandas==2.2.3
requests==2.32.3
tqdm==4.67.1
vk-api==11.9.9
python-dotenv==1.1.0
openpyxl==3.1.5
natasha==1.6.0
spacy==3.8.5
scikit-learn==1.6.1
gigachat==0.1.39.post1
matplotlib==3.10.3
torch==2.7.0
optuna==4.3.0
optuna-integration==4.3.0
catboost==1.2.8
xgboost==3.1.0
